# specs/001-ai-practice-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise AI Intelligent Practice System API
  description: REST API for PPT presentation coaching and sales practice bot
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /practice/history:
    get:
      tags: [practice]
      summary: Get user's practice history
      operationId: getPracticeHistory
      security:
        - BearerAuth: []
      parameters:
        - name: scenario_type
          in: query
          schema:
            type: string
            enum: [presentation, sales]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Practice history
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/PracticeSession'
                  total:
                    type: integer
  /analytics/leaderboard:
    get:
      tags: [analytics]
      summary: Get leaderboard
      operationId: getLeaderboard
      security:
        - BearerAuth: []
      parameters:
        - name: scenario_type
          in: query
          required: true
          schema:
            type: string
            enum: [presentation, sales]
        - name: presentation_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by specific presentation
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
  /analytics/dashboard:
    get:
      tags: [analytics]
      summary: Get admin analytics dashboard
      operationId: getAnalytics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_sessions:
                    type: integer
                  completion_rate:
                    type: number
                    format: float
                  average_score:
                    type: number
                    format: float
                  most_common_gaps:
                    type: array
                    items:
                      type: string
                  score_distribution:
                    type: object
                    properties:
                      excellent:
                        type: integer
                      good:
                        type: integer
                      needs_improvement:
                        type: integer
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
        department:
          type: string
...
    PracticeSession:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        scenario_type:
          type: string
          enum: [presentation, sales]
        status:
          type: string
          enum: [preparing, in_progress, paused, completed, scoring]
        start_time:
          type: string
          format: date-time
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        average_score:
          type: number
          format: float
        total_sessions:
          type: integer
...